+++ imagery/i.points/dot.c
@@ -2,7 +2,7 @@
 #include <grass/raster.h>
 int dot(int x, int y)
 {
-    int vx[5], vy[5];
+    double vx[5], vy[5];
 
     vx[0] = x;
     vy[0] = y - dotsize;
