+++ lib/cairodriver/Draw_line.c
@@ -1,8 +1,8 @@
 #include "cairodriver.h"
 
-void Cairo_draw_line(int x1, int y1, int x2, int y2)
+void Cairo_draw_line(double x1, double y1, double x2, double y2)
 {
-    G_debug(3, "Cairo_draw_line: %d %d %d %d", x1, y1, x2, y2);
+    G_debug(3, "Cairo_draw_line: %f %f %f %f", x1, y1, x2, y2);
 
     if (x1 == x2 && y1 == y2) {
 	/* don't draw degenerate lines */
