+++ lib/driver/Line_width.c
@@ -1,7 +1,7 @@
 #include "driver.h"
 #include "driverlib.h"
 
-void COM_Line_width(int width)
+void COM_Line_width(double width)
 {
     if (driver->Line_width)
 	(*driver->Line_width) (width);
