+++ lib/driver/Set_window.c
@@ -1,7 +1,7 @@
 #include "driver.h"
 #include "driverlib.h"
 
-void COM_Set_window(int t, int b, int l, int r)
+void COM_Set_window(double t, double b, double l, double r)
 {
     if (driver->Set_window)
 	(*driver->Set_window) (t, b, l, r);
