+++ lib/driver/Text_size.c
@@ -1,13 +1,13 @@
 #include "driver.h"
 #include "driverlib.h"
 
-void COM_Text_size(int x, int y)
+void COM_Text_size(double x, double y)
 {
-    text_size_x = (double)x / 25.0;
-    text_size_y = (double)y / 25.0;
+    text_size_x = x / 25.0;
+    text_size_y = y / 25.0;
 }
 
 void COM_Text_rotation(double val)
 {
-    text_rotation = (double)val;
+    text_rotation = val;
 }
