+++ lib/psdriver/Line_width.c
@@ -1,9 +1,9 @@
 #include "psdriver.h"
 
-void PS_Line_width(int width)
+void PS_Line_width(double width)
 {
     if (width < 0)
 	width = 0;
 
-    output("%d WIDTH\n", width);
+    output("%f WIDTH\n", width);
 }
